+++ packages/gallery/src/components/gallery/proGallery/slideshowView.js
@@ -885,7 +885,7 @@ class SlideshowView extends React.Component {
       customNavArrowsRenderer: this.props.customComponents.customNavArrowsRenderer,
       arrowsColor: this.props.options.arrowsColor,
       arrowsSize: this.props.options.arrowsSize,
-      arrowsType: layoutParams.navigationArrows.arrowsType,
+      arrowsType: layoutParams.navigationArrows.type,
     });
 
     const { galleryHeight } = this.props.container;
