+++ packages/lib/src/common/v3DefaultOptions.js
@@ -3,6 +3,9 @@ const v3DefaultOptions = {
     gallerySpacing: 0,
     cropRatio: 1, // determine the ratio of the images when using grid (use 1 for squares grid)
     repeatingGroupTypes: '',
+    navigationArrows: {
+      type: 0,
+    },
   },
   galleryMargin: 0,
   cubeRatio: 1,
