+++ packages/lib/src/settings/options/layoutParams_navigationArrows_type.js
