+++ packages/lib/test/addMigratedOptions.spec.js
@@ -252,6 +252,7 @@ function input() {
     targetItemSize: 650,
     enableScroll: true,
     cropOnlyFill: false,
+    layoutParams_navigationArrows_type: 0,
   };
 }
 
@@ -432,6 +433,7 @@ function expected() {
     layoutParams_navigationArrows_padding: 23,
     layoutParams_navigationArrows_verticalAlignment: 'ITEM_CENTER',
     layoutParams_navigationArrows_size: 23,
+    layoutParams_navigationArrows_type: 0,
     layoutParams_info_layout: 'NO_BACKGROUND',
     layoutParams_info_spacing: 10,
     layoutParams_info_border_width: 0,
